servers/grant-mcp/engine/requirements.txt
pypandoc>=1.12
python-docx>=1.1.0
bibtexparser>=1.4.0
